Moving LinkedIn specific rules into LinkedIn subfolder
