+++ lib/rules/linkedin/no-unguarded-window.js
@@ -1,4 +1,4 @@
-const { isBrowser, isEnvironmentBrowser } = require('../utils/fast-boot');
+const { isBrowser, isEnvironmentBrowser } = require('../utils/bpr');
 
 const MESSAGE = 'Do not use unguarded window references. Please see the following guide for more infromation: http://github.com/chadhietala/ember-best-practices/files/guides/no-unguarded-window.md';
